Plot the indices

Plot the indices

plot_indices(
  data,
  save_loc = getwd(),
  file_name = "index.png",
  legend_loc = "right"
)

Arguments

data

A data frame created by sdmTMB::get_index() with an additional column giving the .id of each index called area. If area is not not present, then all data points will be assumed to be from one default area. It behooves you to make area an ordered factor such that the areas are plotted in the order that you desire.

save_loc

A string providing the location of the directory you wish to save the png file to. The default is your current working directory.

file_name

A string giving the file name. The default is "index.png".

legend_loc

Location for the legend to be added to the figure where the options are "top", "bottom", "left", "right", or a vector of length 2 giving relative coordinate positions. The default is "right".

Value

Creates and saves a plot of the abundance indices in the directory location specified by the save_loc function input.

Details

TODO

Think about adding some plotting information back in to have a title on the figure.

Author

Chantel R. Wetzel