Pull and format data for use in downstream functions.

A list is returned with specifications to run a sdmTMB model, formatted and filtered data, and a prediction grid used to estimate indices of abundance in calc_index_areas() and diagnose().

pull_and_format_data(
  configuration_to_run,
  common_name = NULL,
  survey = "NWFSC.Combo",
  years = 2003:2050,
  verbose = TRUE
)

Arguments

configuration_to_run

A data frame based on the configuration data object that is used to pull and format data list used in downstream functions (e.g., run_sdmtmb()). If a configuration_to_run is specified, the function will pull data from nwfscSurvey::pull_catch() for each line and will create a data list that is used by run_sdmtmb().

common_name

A character string used to pull data from nwfscSurvey::pull_catch(). This is only used if configuration_to_run is not specified.

survey

A character string used to pull data from nwfscSurvey::pull_catch(). Commonly used values are NWFSC.Combo, Triennial, AFSC.Slope, and NWFSC.Slope. This is only used if configuration_to_run is not specified.

years

A numeric vector used to pull data from nwfscSurvey::pull_catch(). This is only used if configuration_to_run is not specified.

verbose

A logical that specifies if you want to print messages and warnings to the console. The default is FALSE.

Value

A list of configurations and data

species

fxn

source

family

formula

min_depth

max_depth

min_latitude

min_year

max_year

anisotropy

knots

spatiotemporal1

spatiotemporal2

share_range

used

data_raw

data_filtred

Author

Chantel R. Wetzel and Kelli F. Johnson